game runs through

+++ db/migrate/20200711170047_create_games.rb
@@ -2,7 +2,7 @@ class CreateGames < ActiveRecord::Migration[6.0]
   def change
     create_table :games do |t|
       t.string :passcode, null: false
-      t.integer :current_castle, default: nil
+      t.float :current_castle, default: nil
       t.references :host, index: true
       t.references :guest, index: true
